+++ tencentcloud/ses/v20201002/client.go
@@ -1633,6 +1633,60 @@ func (c *Client) UpdateEmailIdentityWithContext(ctx context.Context, request *Up
     return
 }
 
+func NewUpdateEmailSmtpPassWordRequest() (request *UpdateEmailSmtpPassWordRequest) {
+    request = &UpdateEmailSmtpPassWordRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    
+    request.Init().WithApiInfo("ses", APIVersion, "UpdateEmailSmtpPassWord")
+    
+    
+    return
+}
+
+func NewUpdateEmailSmtpPassWordResponse() (response *UpdateEmailSmtpPassWordResponse) {
+    response = &UpdateEmailSmtpPassWordResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// UpdateEmailSmtpPassWord
+// 设置邮箱的smtp密码。若要通过smtp发送邮件，必须为邮箱设置smtp密码。初始时，邮箱没有设置smtp密码，不能使用smtp的方式发送邮件。设置smtp密码后，可以修改密码。
+//
+// 可能返回的错误码:
+//  INTERNALERROR = "InternalError"
+//  INVALIDPARAMETERVALUE_INVALIDSMTPPASSWORD = "InvalidParameterValue.InvalidSmtpPassWord"
+//  INVALIDPARAMETERVALUE_NOSUCHSENDER = "InvalidParameterValue.NoSuchSender"
+//  OPERATIONDENIED_REPEATPASSWORD = "OperationDenied.RepeatPassWord"
+func (c *Client) UpdateEmailSmtpPassWord(request *UpdateEmailSmtpPassWordRequest) (response *UpdateEmailSmtpPassWordResponse, err error) {
+    return c.UpdateEmailSmtpPassWordWithContext(context.Background(), request)
+}
+
+// UpdateEmailSmtpPassWord
+// 设置邮箱的smtp密码。若要通过smtp发送邮件，必须为邮箱设置smtp密码。初始时，邮箱没有设置smtp密码，不能使用smtp的方式发送邮件。设置smtp密码后，可以修改密码。
+//
+// 可能返回的错误码:
+//  INTERNALERROR = "InternalError"
+//  INVALIDPARAMETERVALUE_INVALIDSMTPPASSWORD = "InvalidParameterValue.InvalidSmtpPassWord"
+//  INVALIDPARAMETERVALUE_NOSUCHSENDER = "InvalidParameterValue.NoSuchSender"
+//  OPERATIONDENIED_REPEATPASSWORD = "OperationDenied.RepeatPassWord"
+func (c *Client) UpdateEmailSmtpPassWordWithContext(ctx context.Context, request *UpdateEmailSmtpPassWordRequest) (response *UpdateEmailSmtpPassWordResponse, err error) {
+    if request == nil {
+        request = NewUpdateEmailSmtpPassWordRequest()
+    }
+    
+    if c.GetCredential() == nil {
+        return nil, errors.New("UpdateEmailSmtpPassWord require credential")
+    }
+
+    request.SetContext(ctx)
+    
+    response = NewUpdateEmailSmtpPassWordResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewUpdateEmailTemplateRequest() (request *UpdateEmailTemplateRequest) {
     request = &UpdateEmailTemplateRequest{
         BaseRequest: &tchttp.BaseRequest{},