+++ tencentcloud/tione/v20211111/errors.go
@@ -185,6 +185,9 @@ const (
 	// 参数错误。
 	INVALIDPARAMETER = "InvalidParameter"
 
+	// 模型包不符合规范。
+	INVALIDPARAMETER_MODELFILEINVALID = "InvalidParameter.ModelFileInvalid"
+
 	// 无效的接口。
 	INVALIDPARAMETER_TGWINVALIDINTERFACE = "InvalidParameter.TgwInvalidInterface"