+++ tencentcloud/tsf/v20180326/models.go
@@ -2499,6 +2499,7 @@ func (r *CreateClusterRequest) FromJsonString(s string) error {
 // Predefined struct for user
 type CreateClusterResponseParams struct {
 	// 集群ID
+	// 注意：此字段可能返回 null，表示取不到有效值。
 	Result *string `json:"Result,omitempty" name:"Result"`
 
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。