+++ CHANGELOG.md
@@ -1,3 +1,152 @@
+# Release v1.0.561
+
+## T-Sec-DDoS防护(Anti-DDoS)(antiddos) 版本：2020-03-09
+
+### 第 48 次发布
+
+发布时间：2022-12-16 01:05:43
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [SwitchWaterPrintConfig](https://cloud.tencent.com/document/api/1021/59001)
+
+	* 新增入参：CloudSdkProxy
+
+
+修改数据结构：
+
+* [BGPInstance](https://cloud.tencent.com/document/api/1021/57582#BGPInstance)
+
+	* 新增成员：Line, ElasticServiceBandwidth
+
+* [WaterPrintConfig](https://cloud.tencent.com/document/api/1021/57582#WaterPrintConfig)
+
+	* 新增成员：CloudSdkProxy
+
+
+
+
+## 云呼叫中心(ccc) 版本：2020-02-10
+
+### 第 39 次发布
+
+发布时间：2022-12-16 01:10:35
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [HangUpCall](https://cloud.tencent.com/document/api/679/84273)
+
+
+
+## 云服务器(cvm) 版本：2017-03-12
+
+### 第 101 次发布
+
+发布时间：2022-12-16 01:16:54
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [Instance](https://cloud.tencent.com/document/api/213/15753#Instance)
+
+	* 新增成员：DefaultLoginUser, DefaultLoginPort, LatestOperationErrorMsg
+
+
+
+
+## 媒体处理(mps) 版本：2019-06-12
+
+### 第 49 次发布
+
+发布时间：2022-12-16 01:33:15
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [ProcessMedia](https://cloud.tencent.com/document/api/862/37578)
+
+	* 新增入参：TaskType
+
+
+修改数据结构：
+
+* [TaskNotifyConfig](https://cloud.tencent.com/document/api/862/37615#TaskNotifyConfig)
+
+	* <font color="#dd0000">**修改成员**：</font>CmqModel, CmqRegion
+
+
+
+
+## 自然语言处理(nlp) 版本：2019-04-08
+
+### 第 16 次发布
+
+发布时间：2022-12-15 10:27:58
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [GenerateCouplet](https://cloud.tencent.com/document/api/271/84075)
+* [GeneratePoetry](https://cloud.tencent.com/document/api/271/84074)
+
+
+
+## 邮件推送(ses) 版本：2020-10-02
+
+### 第 17 次发布
+
+发布时间：2022-12-16 01:37:37
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [UpdateEmailSmtpPassWord](https://cloud.tencent.com/document/api/1288/84274)
+
+
+
+## 腾讯云 TI 平台 TI-ONE(tione) 版本：2021-11-11
+
+### 第 13 次发布
+
+发布时间：2022-12-16 01:46:34
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateModelService](https://cloud.tencent.com/document/api/851/82291)
+
+	* 新增入参：CallbackUrl
+
+
+
+
+## 腾讯云 TI 平台 TI-ONE(tione) 版本：2019-10-22
+
+
+
 # Release v1.0.560
 
 ## 人体分析(bda) 版本：2020-03-24